\name{opnmfR-package}
\alias{opnmfR-package}
\alias{opnmfR}
\docType{package}
\title{
\packageTitle{opnmfR}
}
\description{
\packageDescription{opnmfR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{opnmfR}
\packageIndices{opnmfR}
~~ OPNMF can be called using opnmfR or opnmfRcpp  ~~
}
\author{
\packageAuthor{opnmfR}

Maintainer: \packageMaintainer{opnmfR}
}
\references{
##Methods
Z. Yang, E. Oja, Linear and nonlinear projective nonnegative matrix factorization, IEEE Trans. Neural Netw., 21 (5) (2010), pp. 734-749

C. Boutsidis, E. Gallopoulos, SVD based initialization: a head start for nonnegative matrix factorization, Pattern Recognit., 41 (4) (2008), pp. 1350-1362

##Applications
*Sotiras, Aristeidis, Jon B. Toledo, Raquel E. Gur, Ruben C. Gur, Theodore D. Satterthwaite, and Christos Davatzikos. Patterns of coordinated cortical remodeling during adolescence and their associations with functional specialization and evolutionary expansion. Proceedings of the National Academy of Sciences 114, no. 13 (2017): 3527-3532.

*Varikuti, Deepthi P., Sarah Genon, Aristeidis Sotiras, Holger Schwender, Felix Hoffstaedter, Kaustubh R. Patil, Christiane Jockwitz, Svenja Caspers, Susanne Moebus, Katrin Amunts, Christos Davatzikos, and Simon Eickhoff. Evaluation of non-negative matrix factorization of grey matter in age prediction. NeuroImage 173 (2018): 394-410.
}
opnmf
nmf
nnmf
matrix factorization

\keyword{ package }
\seealso{
https://github.com/asotiras/brainparts/
}
\examples{
opnmfR_test() 
}
