% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opnmfR.R
\name{opnmfR_ranksel_splithalf}
\alias{opnmfR_ranksel_splithalf}
\title{Rank selection based on split-half similarity measures (cosine, inner-product, cosine 
correlation, and adjusted rand index)}
\usage{
opnmfR_ranksel_splithalf(
  X,
  rs,
  W0 = NULL,
  use.rcpp = TRUE,
  nrepeat = 1,
  similarity = "inner",
  splits = NA,
  plots = TRUE,
  seed = NA,
  rtrue = NA,
  ...
)
}
\arguments{
\item{X}{A matrix, rows are the number of features and columns are number of samples}

\item{rs}{A vector of ranks to test for selection, 
if rs=NULL then \code{1:nrow(X)} is used (default NULL)}

\item{W0}{A string or matrix for initialization (default NULL)}

\item{use.rcpp}{A logical, use \code{opnmfRcpp()} (default TRUE)
dimensional data (default TRUE)}

\item{nrepeat}{A number, number of splits (default 1)}

\item{similarity}{.... (deafault "inner")}

\item{splits}{A list, provided when user wants to define split of sample. Indecies of 
columns for one half of the split and rest are tanken as second split half (default NA)}

\item{plots}{A logical, create a dot plot displaying the similarity measures for 
ranks provided and indicating the highest value for each measure (default TRUE)}

\item{seed}{the \code{set.seed} used to select the split-half (default NA)}

\item{rtrue}{the true rank of the input data if known (default NA)}
}
\value{
A list containing the similarity measures for all methods, the time taken for selection, 
the seed used for split-half, the selected rank based on the highest of each similarity measure
}
\description{
Rank selection based on split-half similarity measures (cosine, inner-product, cosine 
correlation, and adjusted rand index)
}
\examples{
 result <- opnmfR_ranksel_splithalf()
}
