% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opnmfR.R
\name{opnmfR_test_ranksel_synthetic}
\alias{opnmfR_test_ranksel_synthetic}
\title{Rank selection test on synthetic data
Runs all three avilable rank selection methods.}
\usage{
opnmfR_test_ranksel_synthetic(
  n = 100,
  r = 10,
  p = 100,
  rs = NULL,
  W0 = "nndsvd",
  nrepeat = 1
)
}
\arguments{
\item{n}{A number, number of rows (default 100)}

\item{r}{A number, true rank (default 10)}

\item{p}{A number, number of columns (default 100)}

\item{rs}{A vector with  the ranks  you wish to test for selection, 
if rs=NULL then \code{1:min(nrow(X), round(r+r*0.5))} is used (default NULL)}

\item{W0}{A string or matrix for initialization (default "nndsvd")}

\item{nrepeat}{A number, number of iterations for rank selection, 
i.e. number of permutations for \code{opnmfR_ranksel_perm} 
and number of split-halves for \code{opnmfR_ranksel_splithalf} (default 1)}
}
\value{
A list with rank selection outputs from \code{opnmfR_ranksel_perm}, \code{opnmfR_ranksel_ooser}, and
\code{opnmfR_ranksel_splithalf}
}
\description{
Rank selection test on synthetic data
Runs all three avilable rank selection methods.
}
\examples{
result <- opnmfR_test_ranksel_synthetic()
}
\seealso{
\code{\link{opnmfR_ranksel_perm}}, \code{\link{opnmfR_ranksel_ooser}}, and 
\code{\link{opnmfR_ranksel_splithalf}}
}
