% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opnmfR.R
\docType{package}
\name{opnmfR}
\alias{opnmfR}
\title{opnmfR: A package for orthonormal projective non-negative matrix factorization}
\description{
This package provides functions for opnmf factorization, initialization, and 
rank election using R and Rcpp code
}
