% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opnmfR.R
\name{opnmfR_ranksel_perm}
\alias{opnmfR_ranksel_perm}
\title{Permutation based rank selection}
\usage{
opnmfR_ranksel_perm(
  X,
  rs,
  W0 = NULL,
  use.rcpp = TRUE,
  nperm = 1,
  plots = TRUE,
  seed = NA,
  rtrue = NA,
  ...
)
}
\arguments{
\item{X}{A matrix, rows are the number of features and columns are number of samples}

\item{rs}{A vector of ranks to test for selection, 
if rs=NULL then \code{1:nrow(X)} is used (default NULL)}

\item{W0}{A string or matrix for initialization (default NULL)}

\item{use.rcpp}{A logical, use \code{opnmfRcpp()} (default TRUE)
dimensional data (default TRUE)}

\item{nperm}{A number, number of permuatations conducted (default 1)}

\item{plots}{A logical, create a dot plot displaying the similarity measures for 
ranks provided and indicating the highest value for each measure (default TRUE)}

\item{seed}{the \code{set.seed} used to select the split-half (default NA)}

\item{rtrue}{the true rank of the input data if known (default NA)}
}
\value{
A list containing the approximation matrices W and H, of input matrix
X; the interation number at which a solution was found; 
the reconstruction error, time taken for rank selection; the seed used for permutation of 
input data matrix
#' @examples
 result <- opnmfR_ranksel_perm()
}
\description{
Permutation based rank selection
}
